Use central-publishing-maven-plugin for deployment (#51)
* Use central-publishing-maven-plugin for deployment

Since this new plugin doesn't have an option to skip, we reparent
ep-maven and ep-maven-interface under a new module, ep-maven-parent,
which contains the definition. Everything under this parent (Except
itself) will be published to central if the property central-portal
is set (i.e., `-Dcentral-portal=true`). Publishing to central
requires a server with ID `central-portal` to exist in settings.xml with
your access token.

---------

Co-authored-by: Robert Krajewski <rkrajews@tibco.com>